requirements.txt
asgiref
cffi
cryptography
Django
pycparser
PyMySQL
python-dotenv
sqlparse
gunicorn